add email templates, readme

+++ app/Console/Kernel.php
@@ -29,21 +29,31 @@ class Kernel extends ConsoleKernel
     protected function schedule(Schedule $schedule)
     {
         $schedule->call(function () {
-            $current = Carbon::now()->toDateString();
-            $emails = DB::table('email_queue')->where('send_date', $current)->get();
+            $currentStart = Carbon::now()->format('Y-m-d H:i:00');//->toDateTimeString();
+            $currentEnd = Carbon::now()->format('Y-m-d H:i:59');//->toDateTimeString();
+            $emails = DB::table('email_queue')->whereBetween('send_date', [$currentStart,$currentEnd])->get();
+            
             foreach ($emails as $key => $value) {
+                $product = DB::table('products')->find($value->product_id);
                 $email_data = array(
-                    'name' => $value->email
+                    'name' => $product->first_name,
+                    'link' => 'https://id.pheramor.com/status.php?pheramor_id=' . $product->pheramor_id,
+                    'to' => $product->sales_email
                 );
-                if ($value->send_order == 1) {
-                    Mail::to($value->email)->send(new StatusReminder($email_data, 'ship_reminder1'));
-                    DB::table('email_queue')->where('id', $value->id)->delete();
-                } else if($value->send_order == 2) {
-                    Mail::to($value->email)->send(new StatusReminder($email_data, 'ship_reminder2'));
+                // return $this->respond($product);
+                if ($product->swab_returned_date == null) {
+                    if ($value->send_order == 1) {
+                        $this->sendMail($product->sales_email, $product->account_email, $email_data, 'ship_reminder1');
+                        DB::table('email_queue')->where('id', $value->id)->delete();
+                    } else {
+                        $this->sendMail($product->sales_email, $product->account_email, $email_data, 'ship_reminder2');
+                        DB::table('email_queue')->where('id', $value->id)->delete();
+                    }
+                } else {
                     DB::table('email_queue')->where('id', $value->id)->delete();
                 }
             }
-        })->daily();
+        })->everyMinute();
     }
 
     /**
@@ -56,5 +66,18 @@ class Kernel extends ConsoleKernel
         $this->load(__DIR__.'/Commands');
 
         require base_path('routes/console.php');
+    }
+
+    // Send mail
+    public function sendMail($to, $cc, $data, $type) {
+
+        if ($cc == null || $cc == '') {
+            Mail::to($to)->queue(new StatusReminder($data, $type));
+        } else {
+            Mail::to($to)
+                ->cc($cc)
+                ->queue(new StatusReminder($data, $type));
+        }
+
     }
 }
